feat: implemented different input file reading strategy
I have choosed this different approach of file reading because when dealing
with large files it would be impossible to read the whole file at once.
This approach uses an asyncGenerator and reads with fs.readSync() method.
This way the reading process is streamlined. At this point this approach is not tested with large input file.